Repository/code/sentiment-analyzer/requirements.txt
elasticsearch
nltk
ruamel.yaml
textblob